SellerManagement/src/main/java/com/example/sellermanagement/entity/ProductEntry.java
package com.example.sellermanagement.entity;

public class ProductEntry {

    private String productName;
    private double price;
    private String productImage;
    private String category;
    private int quantity;
    private String description;

    public String getProductName() {
        return productName;
    }

    public void setProductName(String productName) {
        this.productName = productName;
    }

    public double getPrice() {
        return price;
    }

    public void setPrice(double price) {
        this.price = price;
    }

    public String getProductImage() {
        return productImage;
    }

    public void setProductImage(String productImage) {
        this.productImage = productImage;
    }

    public String getCategory() {
        return category;
    }

    public void setCategory(String category) {
        this.category = category;
    }

    public int getQuantity() {
        return quantity;
    }

    public void setQuantity(int quantity) {
        this.quantity = quantity;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    @Override
    public String toString() {
        return "ProductEntry{" +
                "productName='" + productName + '\'' +
                ", price=" + price +
                ", productImage='" + productImage + '\'' +
                ", category='" + category + '\'' +
                ", quantity=" + quantity +
                ", description='" + description + '\'' +
                '}';
    }
}
